requirements.txt
aiosql==3.4.1
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
autopep8==1.6.0
cfgv==3.3.1
click==8.1.2
contextlib2==21.6.0
databases==0.5.5
distlib==0.3.4
fastapi==0.75.1
filelock==3.6.0
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
identify==2.4.12
idna==3.3
loguru==0.6.0
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
nodeenv==1.6.0
platformdirs==2.5.1
pre-commit==2.18.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
PyJWT==2.3.0
PyPika==0.48.9
python-dotenv==0.20.0
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.34
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.6
virtualenv==20.14.0
